requirements.txt
fastapi
requests
uvicorn
python-dotenv
openai>=1.0.0